Removed triple quotes for comments
According to this link; Python only has one way of doing comments and that is using #
Triple quotes are treated as regular strings with the exception that they can span multiple lines. By regular strings I mean that if they are not assigned to a variable they will be immediately garbage collected as soon as that code executes. hence are not ignored by the interpreter in the same way that #a comment is.
https://www.codecademy.com/forum_questions/505ba3cfc6addb000200e33c#:~:text=As%20part%20of%20the%20Python,they%20can%20span%20multiple%20lines.